finalized func libraries
